+++ testnet/stacks-node/src/tests/neon_integrations.rs
@@ -71,10 +71,6 @@ use super::{
     make_microblock, make_stacks_transfer, make_stacks_transfer_mblock_only, to_addr, ADDR_4, SK_1,
     SK_2,
 };
-use stacks::core::StacksEpoch;
-use stacks::core::StacksEpochId;
-use stacks::vm::ContractName;
-use std::convert::TryFrom;
 
 pub fn neon_integration_test_conf() -> (Config, StacksAddress) {
     let mut conf = super::new_test_conf();
@@ -5818,314 +5814,4 @@ fn atlas_stress_integration_test() {
     }
 
     test_observer::clear();
-}
-
-/// Deserializes the `StacksTransaction` objects from `blocks` and returns all those that
-/// match `test_fn`.
-fn select_transactions_where(
-    blocks: &Vec<serde_json::Value>,
-    test_fn: fn(&StacksTransaction) -> bool,
-) -> Vec<StacksTransaction> {
-    let mut result = vec![];
-    for block in blocks {
-        let transactions = block.get("transactions").unwrap().as_array().unwrap();
-        for tx in transactions.iter() {
-            let raw_tx = tx.get("raw_tx").unwrap().as_str().unwrap();
-            let tx_bytes = hex_bytes(&raw_tx[2..]).unwrap();
-            let parsed = StacksTransaction::consensus_deserialize(&mut &tx_bytes[..]).unwrap();
-            if test_fn(&parsed) {
-                eprintln!("show-block {:#?}", &block);
-                eprintln!("show-parsed {:#?}", &parsed);
-                result.push(parsed);
-            }
-        }
-    }
-
-    return result;
-}
-
-/// This function will call `next_block_and_wait` until the burnchain height underlying `BitcoinRegtestController`
-/// reaches *at least* `target_height`.
-///
-/// Returns `false` if `next_block_and_wait` times out.
-///
-/// Note: The user cannot assume that the block height reached is *exactly* `target_height`. It
-/// seems that the chain may often advanced `1` block past `target_height`.
-fn run_until_burnchain_height(
-    btc_regtest_controller: &mut BitcoinRegtestController,
-    blocks_processed: &Arc<AtomicU64>,
-    burnchain_height: &Arc<AtomicU64>,
-    target_height: u64,
-) -> bool {
-    let current = burnchain_height.load(Ordering::SeqCst);
-    eprintln!(
-        "run_until_burnchain_height: Issuing block at {}, current burnchain height is ({})",
-        get_epoch_time_secs(),
-        current
-    );
-
-    next_block_and_wait(btc_regtest_controller, &blocks_processed);
-    while burnchain_height.load(Ordering::SeqCst) <= target_height {
-        let current = burnchain_height.load(Ordering::SeqCst);
-        eprintln!(
-            "run_until_burnchain_height: Issuing block at {}, current burnchain height is ({})",
-            get_epoch_time_secs(),
-            current
-        );
-        let next_result = next_block_and_wait(btc_regtest_controller, &blocks_processed);
-        if !next_result {
-            return false;
-        }
-    }
-
-    true
-}
-
-/// This test checks that the block limit is changed at Stacks 2.05. We lower the allowance, and
-/// check that we can 1) afford the function call before the target height and
-/// 2) cannot afford the function call after the target height.
-#[test]
-#[ignore]
-fn test_cost_limit_switch_version205() {
-    if env::var("BITCOIND_TEST") != Ok("1".into()) {
-        return;
-    }
-
-    // This contract contains `increment-many`, which does many MARF reads.
-    let giant_contract = r#"
-;; define counter variable
-(define-data-var counter int 0)
-
-;; increment method
-(define-public (increment)
-  (begin
-    (var-set counter (+ (var-get counter) 1))
-    (ok (var-get counter))))
-
-  (define-public (increment-many (unused-index uint))
-    (begin
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (unwrap! (increment) (err u1))
-      (ok (var-get counter))))
-    "#
-    .to_string();
-
-    // Create three characters, `creator`, `alice` and `bob`.
-    let creator_sk = StacksPrivateKey::new();
-    let creator_addr = to_addr(&creator_sk);
-    let creator_pd: PrincipalData = creator_addr.into();
-
-    let alice_sk = StacksPrivateKey::new();
-    let alice_addr = to_addr(&alice_sk);
-    let alice_pd: PrincipalData = alice_addr.into();
-
-    let bob_sk = StacksPrivateKey::new();
-    let bob_addr = to_addr(&bob_sk);
-    let bob_pd: PrincipalData = bob_addr.into();
-
-    let (mut conf, _) = neon_integration_test_conf();
-
-    // Create a schedule where we lower the read_count on Epoch2_05.
-    conf.burnchain.epochs = Some(vec![
-        StacksEpoch {
-            epoch_id: StacksEpochId::Epoch20,
-            start_height: 0,
-            end_height: 220,
-            block_limit: ExecutionCost {
-                write_length: 100000000,
-                write_count: 1000,
-                read_length: 1000000000,
-                read_count: 150,
-                runtime: 5000000000,
-            },
-        },
-        StacksEpoch {
-            epoch_id: StacksEpochId::Epoch2_05,
-            start_height: 220,
-            end_height: 9223372036854775807,
-            block_limit: ExecutionCost {
-                write_length: 100000000,
-                write_count: 1000,
-                read_length: 1000000000,
-                read_count: 50,
-                runtime: 5000000000,
-            },
-        },
-    ]);
-
-    conf.initial_balances.push(InitialBalance {
-        address: alice_pd.clone(),
-        amount: 10492300000,
-    });
-    conf.initial_balances.push(InitialBalance {
-        address: bob_pd.clone(),
-        amount: 10492300000,
-    });
-    conf.initial_balances.push(InitialBalance {
-        address: creator_pd.clone(),
-        amount: 10492300000,
-    });
-
-    test_observer::spawn();
-    conf.events_observers.push(EventObserverConfig {
-        endpoint: format!("localhost:{}", test_observer::EVENT_OBSERVER_PORT),
-        events_keys: vec![EventKeyType::AnyEvent],
-    });
-
-    let mut btcd_controller = BitcoinCoreController::new(conf.clone());
-    btcd_controller
-        .start_bitcoind()
-        .map_err(|_e| ())
-        .expect("Failed starting bitcoind");
-
-    let mut btc_regtest_controller = BitcoinRegtestController::new(conf.clone(), None);
-    let http_origin = format!("http://{}", &conf.node.rpc_bind);
-
-    btc_regtest_controller.bootstrap_chain(200);
-
-    let mut run_loop = neon::RunLoop::new(conf);
-    let blocks_processed = run_loop.get_blocks_processed_arc();
-    let burnchain_height = run_loop.get_burnchain_height_arc();
-
-    let channel = run_loop.get_coordinator_channel().unwrap();
-
-    thread::spawn(move || run_loop.start(None, 0));
-
-    // Wait until block 210, so we now that the burnchain is ready.
-    wait_for_runloop(&blocks_processed);
-    run_until_burnchain_height(
-        &mut btc_regtest_controller,
-        &blocks_processed,
-        &burnchain_height,
-        210,
-    );
-
-    // Publish the contract so we can use it.
-    submit_tx(
-        &http_origin,
-        &make_contract_publish(
-            &creator_sk,
-            0,
-            1100000,
-            "increment-contract",
-            &giant_contract,
-        ),
-    );
-
-    // Wait to make sure the contract is published.
-    run_until_burnchain_height(
-        &mut btc_regtest_controller,
-        &blocks_processed,
-        &burnchain_height,
-        215,
-    );
-
-    // Check that we have defined the contract.
-    let increment_contract_defines = select_transactions_where(
-        &test_observer::get_blocks(),
-        |transaction| match &transaction.payload {
-            TransactionPayload::SmartContract(contract) => {
-                contract.name == ContractName::try_from("increment-contract").unwrap()
-            }
-            _ => false,
-        },
-    );
-    assert_eq!(increment_contract_defines.len(), 1);
-
-    // Alice calls the contract and should succeed, because we have not lowered the block limit
-    // yet.
-    submit_tx(
-        &http_origin,
-        &make_contract_call(
-            &alice_sk,
-            0,
-            1000,
-            &creator_addr.into(),
-            "increment-contract",
-            "increment-many",
-            &[Value::UInt(0 as u128)],
-        ),
-    );
-
-    // Wait for the contract call to process.
-    run_until_burnchain_height(
-        &mut btc_regtest_controller,
-        &blocks_processed,
-        &burnchain_height,
-        220,
-    );
-
-    // Check that we have processed the contract successfully, by checking that the contract call
-    // is in the block record.
-    let increment_calls_alice = select_transactions_where(
-        &test_observer::get_blocks(),
-        |transaction| match &transaction.payload {
-            TransactionPayload::ContractCall(contract) => {
-                contract.contract_name == ContractName::try_from("increment-contract").unwrap()
-            }
-            _ => false,
-        },
-    );
-    assert_eq!(increment_calls_alice.len(), 1);
-
-    // Clear the observer so we can look for Bob's transaction.
-    test_observer::clear();
-
-    // The cost contract was switched at height 220. So, now we expect Bob's call to fail.
-    run_until_burnchain_height(
-        &mut btc_regtest_controller,
-        &blocks_processed,
-        &burnchain_height,
-        230,
-    );
-    submit_tx(
-        &http_origin,
-        &make_contract_call(
-            &bob_sk,
-            0,
-            1000,
-            &creator_addr.into(),
-            "increment-contract",
-            "increment-many",
-            &[Value::UInt(0 as u128)],
-        ),
-    );
-
-    // Wait for the contract to finish.
-    run_until_burnchain_height(
-        &mut btc_regtest_controller,
-        &blocks_processed,
-        &burnchain_height,
-        235,
-    );
-
-    // Bob's calls didn't work because he called after the block limit was lowered.
-    let increment_calls_bob = select_transactions_where(
-        &test_observer::get_blocks(),
-        |transaction| match &transaction.payload {
-            TransactionPayload::ContractCall(contract) => {
-                contract.contract_name == ContractName::try_from("increment-contract").unwrap()
-            }
-            _ => false,
-        },
-    );
-    assert_eq!(increment_calls_bob.len(), 0);
-
-    channel.stop_chains_coordinator();
 }
